PaperAnalysis/requirements.txt
deepeval == 2.3.3
docling == 2.28.4
docling-core == 2.26.0
python-dotenv == 1.0.1
langchain == 0.3.22
langchain_core == 0.3.50
pandas == 2.2.3
pillow == 11.1.0
ProtoLLM == 0.1.2
pydantic == 2.9.2
dotenv == 0.9.9
open-clip-torch == 2.32.0
sentence_transformers == 4.1.0
chromadb == 1.0.6